Also move PackEmitter
