+++ jflex/src/main/java/jflex/generator/HiLowEmitter.java
@@ -9,8 +9,6 @@
 
 package jflex.generator;
 
-import jflex.core.PackEmitter;
-
 /**
  * HiLowEmitter
  *
