+++ jflex/src/test/java/jflex/core/BUILD
@@ -8,27 +8,11 @@ java_test(
 )
 
 java_test(
-    name = "EmitterTest",
-    srcs = ["EmitterTest.java"],
-    deps = [
-        "//jflex/src/main/java/jflex/generator",
-    ],
-)
-
-java_test(
     name = "IntCharSetTest",
     srcs = ["IntCharSetTest.java"],
     deps = [
         "//jflex/src/main/java/jflex/core",
         "//jflex/src/main/java/jflex/chars",
-    ],
-)
-
-java_test(
-    name = "PackEmitterTest",
-    srcs = ["PackEmitterTest.java"],
-    deps = [
-        "//jflex/src/main/java/jflex/core",
     ],
 )
