+++ jflex/src/test/java/jflex/generator/BUILD
@@ -1,0 +1,15 @@
+java_test(
+    name = "EmitterTest",
+    srcs = ["EmitterTest.java"],
+    deps = [
+        "//jflex/src/main/java/jflex/generator",
+    ],
+)
+
+java_test(
+    name = "PackEmitterTest",
+    srcs = ["PackEmitterTest.java"],
+    deps = [
+        "//jflex/src/main/java/jflex/generator",
+    ],
+)
