+++ jflex/src/test/java/jflex/generator/EmitterTest.java
@@ -7,12 +7,11 @@
  *                                                                         *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-package jflex.core;
+package jflex.generator;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import jflex.generator.Emitter;
 import junit.framework.TestCase;
 
 /**
